fix: align trim/parse_int signatures with callers, fix level_t::level type

+++ test/collezioni_polimorfiche_e_dove_trovarle/dod.cpp
@@ -24,7 +24,7 @@ struct on_off_t
 
 struct level_t // cppcheck-suppress ctuOneDefinitionRuleViolation
 {
-  size_t level;
+  int level;
 };
 
 struct on_off_light_t // cppcheck-suppress ctuOneDefinitionRuleViolation
